Restrict usage to single mail and message template sends
This commit updates the extension to ensure that it only alters
the mail parameters of outgoing emails that are either single emails
(such as those sent manually from a contact record) and template-driven
emails (such as event registration confirmation emails).

Inconsistent values in the $params payload passed through to the hook
mean that immediate usage for CiviMail and CiviMosaico sends is
problematic. As those functions make use of From: addresses defined
under system settings, there is less concern that they will create
messages attempting to use an invalid From: address.